# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mathlers API — boxing-themed AI math competition platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RecordCard {
  id: number;
  title: string;
  /** @nullable */
  holder?: string | null;
  value: number;
  unit: string;
  /** @nullable */
  previousValue?: number | null;
  category: string;
  /** @nullable */
  subcategory?: string | null;
  difficulty: string;
  ageRange?: string;
  /** @nullable */
  recordDate?: string | null;
  /** @nullable */
  location?: string | null;
  source: string;
  /** @nullable */
  sourceUrl?: string | null;
  verified: boolean;
  version?: number;
  createdAt: string;
}

export interface RecordCardInput {
  title: string;
  holder?: string;
  value: number;
  unit: string;
  previousValue?: number;
  category: string;
  subcategory?: string;
  difficulty: string;
  ageRange?: string;
  recordDate?: string;
  location?: string;
  source: string;
  sourceUrl?: string;
}

export type RecordCardStatsByCategoryItem = {
  category: string;
  count: number;
};

export interface RecordCardStats {
  total: number;
  byCategory: RecordCardStatsByCategoryItem[];
  recentlyAdded: number;
}

export interface ScrapeInput {
  sources: string[];
  categories?: string[];
}

export interface ScrapeResult {
  jobId: string;
  status: string;
  message: string;
}

export interface QuestionTemplate {
  id: number;
  name: string;
  topic: string;
  difficulty: string;
  templateText: string;
  variables: string[];
  mathOperation: string;
  /** @nullable */
  exampleOutput?: string | null;
  usageCount?: number;
  createdAt: string;
}

export interface QuestionTemplateInput {
  name: string;
  topic: string;
  difficulty: string;
  templateText: string;
  variables: string[];
  mathOperation: string;
  exampleOutput?: string;
}

export interface Question {
  id: number;
  recordCardId: number;
  templateId: number;
  roundName: string;
  topic: string;
  difficulty: string;
  scenario: string;
  questionText: string;
  options: string[];
  correctAnswer: string;
  steps: string[];
  /** @nullable */
  knockoutChallenge?: string | null;
  hash: string;
  validated: boolean;
  timeLimitSeconds?: number;
  pointValue?: number;
  createdAt: string;
  recordCard?: RecordCard;
}

export interface GenerateQuestionInput {
  recordCardId: number;
  difficulty: string;
  topic: string;
  /** @nullable */
  templateId?: number | null;
  roundName?: string;
  /** @nullable */
  providerId?: string | null;
}

export interface ValidationResult {
  valid: boolean;
  message: string;
  /** @nullable */
  recomputedAnswer?: string | null;
  steps?: string[];
}

export interface Student {
  id: number;
  username: string;
  displayName: string;
  /** @nullable */
  email?: string | null;
  ageGroup: string;
  /** @nullable */
  school?: string | null;
  elo: number;
  belt: string;
  streak: number;
  totalMatches: number;
  totalWins: number;
  favoriteTopics?: string[];
  createdAt: string;
}

export interface StudentInput {
  username: string;
  displayName: string;
  email?: string;
  ageGroup: string;
  school?: string;
  favoriteTopics?: string[];
}

export interface StudentStats {
  studentId: number;
  totalQuestions: number;
  correctAnswers: number;
  accuracy: number;
  avgResponseTime: number;
  currentStreak: number;
  longestStreak: number;
  elo: number;
  belt: string;
  weeklyGrowth?: number;
}

export interface TopicMastery {
  topic: string;
  mastery: number;
  accuracy: number;
  avgResponseTime: number;
  questionsAttempted: number;
  trend: string;
}

export interface Recommendations {
  practiceTopics: string[];
  nextDifficulty: string;
  dailyChallenge: Question;
  weakAreas?: string[];
}

export interface RoundSummary {
  roundNumber: number;
  roundName: string;
  questionId: number;
  correct: boolean;
  responseTime: number;
  pointsEarned: number;
  selectedAnswer?: string;
}

export interface Match {
  id: number;
  studentId: number;
  mode: string;
  difficulty: string;
  status: string;
  currentRound: number;
  totalRounds?: number;
  score: number;
  accuracy: number;
  streak: number;
  comboMultiplier?: number;
  currentQuestion?: Question;
  rounds?: RoundSummary[];
  createdAt: string;
}

export interface MatchInput {
  studentId: number;
  mode: string;
  difficulty: string;
  topics?: string[];
  /** @nullable */
  opponentId?: number | null;
}

export interface AnswerInput {
  questionId: number;
  selectedAnswer: string;
  responseTimeMs: number;
}

export interface AnswerResult {
  correct: boolean;
  correctAnswer: string;
  pointsEarned: number;
  comboMultiplier?: number;
  explanation: string;
  steps?: string[];
  matchComplete: boolean;
  nextQuestion: Question;
}

export interface MatchSummary {
  matchId: number;
  score: number;
  accuracy: number;
  belt: string;
  eloChange: number;
  newElo: number;
  knockouts?: number;
  longestCombo?: number;
  fastestAnswer?: number;
  rounds: RoundSummary[];
}

export interface LeaderboardEntry {
  rank: number;
  studentId: number;
  username: string;
  displayName: string;
  /** @nullable */
  school?: string | null;
  belt: string;
  elo: number;
  winRate: number;
  totalMatches: number;
  streak?: number;
}

export interface AnalyticsOverview {
  totalStudents: number;
  totalMatches: number;
  totalQuestions: number;
  totalRecordCards: number;
  avgAccuracy: number;
  activeToday: number;
  questionsGeneratedToday?: number;
  topCategory?: string;
}

export interface DifficultyBucket {
  difficulty: string;
  topic: string;
  count: number;
  avgAccuracy: number;
}

export interface TrendPoint {
  date: string;
  matchesPlayed: number;
  avgScore: number;
  avgAccuracy: number;
  newStudents?: number;
}

export interface TopicBreakdown {
  topic: string;
  totalAttempts: number;
  correctAttempts: number;
  accuracy: number;
  avgTime: number;
}

export interface Provider {
  id: string;
  name: string;
  status: string;
  model: string;
  requestCount: number;
  errorRate: number;
  avgLatencyMs?: number;
  rateLimitRpm?: number;
}

export interface ProviderUsage {
  totalRequests: number;
  successfulRequests: number;
  failedRequests: number;
  providers: Provider[];
}

export type DashboardDataUpcomingEventsItem = {
  id: number;
  name: string;
  startAt: string;
  type: string;
  participants?: number;
};

export interface DashboardData {
  student: Student;
  stats: StudentStats;
  mastery: TopicMastery[];
  recentMatches: Match[];
  upcomingEvents: DashboardDataUpcomingEventsItem[];
  recommendations: Recommendations;
}

export type ListRecordCardsParams = {
category?: string;
difficulty?: string;
limit?: number;
offset?: number;
};

export type ListTemplatesParams = {
topic?: string;
difficulty?: string;
};

export type ListQuestionsParams = {
difficulty?: string;
topic?: string;
recordCardId?: number;
limit?: number;
offset?: number;
};

export type ListStudentsParams = {
limit?: number;
offset?: number;
};

export type ListMatchesParams = {
studentId?: number;
status?: string;
limit?: number;
offset?: number;
};

export type GetLeaderboardParams = {
period?: GetLeaderboardPeriod;
scope?: GetLeaderboardScope;
topic?: string;
limit?: number;
offset?: number;
};

export type GetLeaderboardPeriod = typeof GetLeaderboardPeriod[keyof typeof GetLeaderboardPeriod];


export const GetLeaderboardPeriod = {
  daily: 'daily',
  weekly: 'weekly',
  monthly: 'monthly',
  'all-time': 'all-time',
} as const;

export type GetLeaderboardScope = typeof GetLeaderboardScope[keyof typeof GetLeaderboardScope];


export const GetLeaderboardScope = {
  global: 'global',
  school: 'school',
  class: 'class',
  'age-group': 'age-group',
} as const;

export type GetPerformanceTrendParams = {
days?: number;
};

export type GetDashboardParams = {
studentId?: number;
};