# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Mathlers API — boxing-themed AI math competition platform
servers:
  - url: /api/v1
    description: Versioned API — v1
  - url: /api
    description: Legacy base path (backward-compat, deprecated)
...
paths:
...
  /questions/generate:
    post:
      operationId: generateQuestion
      tags: [questions]
      summary: Generate a new question using AI pipeline
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateQuestionInput"
      responses:
        "200":
          description: Generated question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
...
  /matches:
    get:
      operationId: listMatches
      tags: [matches]
      summary: List matches
      parameters:
        - in: query
          name: studentId
          schema: { type: integer }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
        - in: query
          name: offset
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: List of matches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Match"
    post:
      operationId: createMatch
      tags: [matches]
      summary: Start a new match
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatchInput"
      responses:
        "201":
          description: Match created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
  /matches/{id}:
    get:
      operationId: getMatch
      tags: [matches]
      summary: Get a match by ID
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Match
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
  /matches/{id}/answer:
    post:
      operationId: submitAnswer
      tags: [matches]
      summary: Submit an answer for the current round
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnswerInput"
      responses:
        "200":
          description: Answer result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnswerResult"
  /matches/{id}/complete:
    post:
      operationId: completeMatch
      tags: [matches]
      summary: Complete a match and save results
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Match summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchSummary"

  # ─── LEADERBOARDS ───────────────────────────────────────────────
...
components:
  schemas:
...
    RecordCard:
      type: object
      required: [id, title, value, unit, category, difficulty, source, verified, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        holder: { type: ["string", "null"] }
        value: { type: number }
        unit: { type: string }
        previousValue: { type: ["number", "null"] }
        category: { type: string }
        subcategory: { type: ["string", "null"] }
        difficulty: { type: string }
        ageRange: { type: string }
        recordDate: { type: ["string", "null"] }
        location: { type: ["string", "null"] }
        source: { type: string }
        sourceUrl: { type: ["string", "null"] }
        verified: { type: boolean }
        version: { type: integer }
        createdAt: { type: string }
...
    Question:
      type: object
      required: [id, recordCardId, templateId, roundName, topic, difficulty, scenario, questionText, options, correctAnswer, steps, hash, validated, createdAt]
      properties:
        id: { type: integer }
        recordCardId: { type: integer }
        templateId: { type: integer }
        roundName: { type: string }
        topic: { type: string }
        difficulty: { type: string }
        scenario: { type: string }
        questionText: { type: string }
        options:
          type: array
          items: { type: string }
        correctAnswer: { type: string }
        steps:
          type: array
          items: { type: string }
        knockoutChallenge: { type: ["string", "null"] }
        hash: { type: string }
        validated: { type: boolean }
        timeLimitSeconds: { type: integer }
        pointValue: { type: integer }
        createdAt: { type: string }
        recordCard:
          $ref: "#/components/schemas/RecordCard"
    GenerateQuestionInput:
      type: object
      required: [recordCardId, difficulty, topic]
      properties:
        recordCardId: { type: integer }
        difficulty: { type: string }
        topic: { type: string }
        templateId: { type: ["integer", "null"] }
        roundName: { type: string }
        providerId: { type: ["string", "null"] }
...
    Match:
      type: object
      required: [id, studentId, mode, difficulty, status, currentRound, score, accuracy, streak, createdAt]
      properties:
        id: { type: integer }
        studentId: { type: integer }
        mode: { type: string }
        difficulty: { type: string }
        status: { type: string }
        currentRound: { type: integer }
        totalRounds: { type: integer }
        score: { type: integer }
        accuracy: { type: number }
        streak: { type: integer }
        comboMultiplier: { type: number }
        currentQuestion:
          $ref: "#/components/schemas/Question"
        rounds:
          type: array
          items:
            $ref: "#/components/schemas/RoundSummary"
        createdAt: { type: string }
    MatchInput:
      type: object
      required: [studentId, mode, difficulty]
      properties:
        studentId: { type: integer }
        mode: { type: string }
        difficulty: { type: string }
        topics:
          type: array
          items: { type: string }
        opponentId: { type: ["integer", "null"] }
    RoundSummary:
      type: object
      required: [roundNumber, roundName, questionId, correct, responseTime, pointsEarned]
      properties:
        roundNumber: { type: integer }
        roundName: { type: string }
        questionId: { type: integer }
        correct: { type: boolean }
        responseTime: { type: number }
        pointsEarned: { type: integer }
        selectedAnswer: { type: string }
    AnswerInput:
      type: object
      required: [questionId, selectedAnswer, responseTimeMs]
      properties:
        questionId: { type: integer }
        selectedAnswer: { type: string }
        responseTimeMs: { type: integer }
    AnswerResult:
      type: object
      required: [correct, correctAnswer, pointsEarned, explanation, matchComplete, nextQuestion]
      properties:
        correct: { type: boolean }
        correctAnswer: { type: string }
        pointsEarned: { type: integer }
        comboMultiplier: { type: number }
        explanation: { type: string }
        steps:
          type: array
          items: { type: string }
        matchComplete: { type: boolean }
        nextQuestion:
          $ref: "#/components/schemas/Question"
    MatchSummary:
      type: object
      required: [matchId, score, accuracy, belt, eloChange, newElo, rounds]
      properties:
        matchId: { type: integer }
        score: { type: integer }
        accuracy: { type: number }
        belt: { type: string }
        eloChange: { type: integer }
        newElo: { type: integer }
        knockouts: { type: integer }
        longestCombo: { type: integer }
        fastestAnswer: { type: number }
        rounds:
          type: array
          items:
            $ref: "#/components/schemas/RoundSummary"
...