# lib/api-zod/src/generated/types/matchSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mathlers API — boxing-themed AI math competition platform
 * OpenAPI spec version: 0.1.0
 */
import type { RoundSummary } from './roundSummary';

export interface MatchSummary {
  matchId: number;
  score: number;
  accuracy: number;
  belt: string;
  eloChange: number;
  newElo: number;
  knockouts?: number;
  longestCombo?: number;
  fastestAnswer?: number;
  rounds: RoundSummary[];
}